pyproject.toml
[tool.poetry]
name = "transparency-bot"
version = "0.1.0"
description = ""
authors = ["idmyn <hello@davidmyno.rs>"]

[tool.poetry.scripts]
bot = "transparency_bot.bot:run"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^7.0.0"
python-telegram-bot = "^12.4.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
